JWST script for working with 3D datacubes
This rough version reads in a MIRI FITS file, and performs tasks such
as subtracting the first frame from all concequent, calculated the
slope for all pixels over all frames, generates a slope FITS image, and
finally prints out the median, average, and standard deviation of the
slope image in counts/frame.

Spet. 10 - Getting erroneous values for the median, average, and stand.
dev., compared to collegues. Does this have something to do with the
way data is handled w/ MRDFITS vs. READFITS? (One reads in values and
UINT, while the other reads in as INT).