Updated version.
This version calculates the median, average, and standard deviation of
the the slope image in counts/sec. The function used to calculate the
slope of each pixel is in terms of frames (x-axis) and counts (y-axis).

Sept 10: This version is still rough, and needs to be cleaned up;
mostly test lines need to be removed and more concise documentation.
This script might become useful for NIRSpec datacubes as well.