Photometry training scripts
First attempt at automating DAOPHOT daofind and phot on images in a
directory. Able to use the counts FITS images it finds (*_cts.fits) and
conduct the daofind task on them.

Next step is to include the phot taks on all the images, then perform a
text dump of the necessary data. Finally, I also plan to add a task to
combine all data into one table, for use in constructing a useful CMD.